+++ src/ficheros/fichero2.txt
@@ -1,0 +1,10 @@
+atun
+salado
+leche
+rancia
+bonsai
+de
+la
+risa
+salmon
+crudo
